Clean up and document functions
